+++ frontend/src/themes/index.ts
@@ -5,7 +5,10 @@ import { catppuccinTheme } from './catppuccinTheme'
 import { scienceTheme } from './scienceTheme'
 import { gruvboxTheme } from './gruvboxTheme'
 import { solarizedTheme } from './solarizedTheme'
+import { kanagawaTheme } from './kanagawaTheme'
+import { kanagawaWaveTheme } from './kanagawaWaveTheme'
+import { draculaTheme } from './draculaTheme'
 
-export { defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme, gruvboxTheme, solarizedTheme }
+export { defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme, gruvboxTheme, solarizedTheme, kanagawaTheme, kanagawaWaveTheme, draculaTheme }
 
-export const themes = [defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme, gruvboxTheme, solarizedTheme]
+export const themes = [defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme, gruvboxTheme, solarizedTheme, kanagawaTheme, kanagawaWaveTheme, draculaTheme]